+++ src/app_manager/consts.py
@@ -1,0 +1,6 @@
+# stopping conditions mapping
+STOPPING_CONDITIONS = {
+    "time": "TIME_STOPPING_CONDITION",
+    "fitness": "FITNESS_STOPPING_CONDITION",
+    "generations": "GENERATIONS_STOPPING_CONDITION"
+}