+++ src/app_manager/problem.py
@@ -1,4 +1,3 @@
-import ujson as json
 from dataclasses import dataclass, field, asdict
 from datetime import datetime